+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/past/emitter/Poet.java
@@ -14,6 +14,7 @@ import org.openprovenance.prov.template.compiler.past.type.ParameterizedType;
 import org.openprovenance.prov.template.compiler.past.type.TypeVariable;
 
 import java.util.*;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 import static org.openprovenance.prov.template.compiler.past.BinaryOp.INSTANCEOF;
@@ -41,7 +42,7 @@ public class Poet implements Emitter<TypeSpec> {
 
         TypeSpec.Builder builder=(clazz.isInterface)?TypeSpec.interfaceBuilder(clazz.name):TypeSpec.classBuilder(clazz.name);
         for (TypeVariable tv : clazz.typeVariables) {
-            builder.addTypeVariable(TypeVariableName.get(tv.name));
+            builder.addTypeVariable(convertTypeVariable(tv));
         }
         clazz.modifiers.forEach(builder::addModifiers);
         // emit superclass if present
@@ -75,7 +76,7 @@ public class Poet implements Emitter<TypeSpec> {
               ;
 
         for (TypeVariable tv : clazz.typeVariables) {
-            builder.addTypeVariable(TypeVariableName.get(tv.name));
+            builder.addTypeVariable(convertTypeVariable(tv));
         }
         clazz.modifiers.forEach(builder::addModifiers);
         clazz.fields.forEach(field -> builder.addField(convert(field)));
@@ -675,7 +676,12 @@ public class Poet implements Emitter<TypeSpec> {
                 CodeBlock argsCode = CodeBlock.join(
                         methodCall.arguments.stream().map(this::convert).collect(Collectors.toList()),
                         ",");
-                return CodeBlock.of("super.$L($L)", methodCall.methodName, argsCode);
+
+                if (methodCall.methodName==null) {
+                    return CodeBlock.of("super($L)", argsCode);
+                } else {
+                    return CodeBlock.of("super.$L($L)", methodCall.methodName, argsCode);
+                }
             }
             case OBJECT_ACCESSOR -> {
                 if (methodCall.className != null) {
@@ -722,7 +728,7 @@ public class Poet implements Emitter<TypeSpec> {
             }
             case VARIABLE -> {
                 TypeVariable tn = (TypeVariable) typeName;
-                return TypeVariableName.get(tn.name);
+                return convertTypeVariable(tn);
             }
             case ARRAY -> {
                 ArrayType at = (ArrayType) typeName;
@@ -735,6 +741,14 @@ public class Poet implements Emitter<TypeSpec> {
                         Arrays.stream(pt.typeArguments).map(this::convert).toArray(TypeName[]::new));
             }
             default -> throw new IllegalArgumentException("conversion not supported yet");
+        }
+    }
+
+    private TypeVariableName convertTypeVariable(TypeVariable tn) {
+        if (tn.bounds.isEmpty()) {
+            return TypeVariableName.get(tn.name);
+        } else {
+            return TypeVariableName.get(tn.name).withBounds(tn.bounds.stream().map(this::convert).toArray(TypeName[]::new));
         }
     }
 
@@ -761,6 +775,7 @@ public class Poet implements Emitter<TypeSpec> {
                     case "int" ->  { return TypeName.get(int.class); }
                     case "int[]" ->  { return ArrayTypeName.of(TypeName.get(int.class)); }
                     case "System" ->  { return ClassName.get(System.class); }
+                    case "AtomicInteger" ->  { return TypeName.get(AtomicInteger.class); }
                     default ->  { /* continue */ }
                 }
             }